// src/types/alignment-line.ts
/**
 * Представляет символ и его измеренную ширину в пикселях.
 */
export interface CharInfo {
  /**
   * Символ.
   */
  readonly char: string;
  /**
   * Ширина символа.
   */
  readonly width: number;
}

/**
 * Представляет пару строк для визуального выравнивания верхнюю строку и соответствующую ей нижнюю строку.
 */
export interface LinePair {
  /**
   * Массив символов верхней последовательности с их ширинами.
   */
  readonly top: CharInfo[];
  /**
   * Массив символов нижней последовательности с их ширинами.
   */
  readonly bottom: CharInfo[];
}